Check for setup/teardown calls at compile time.
Classes that contain at least one `@BindPref` annotation
are now checked at compile-time to properly call
PreferenceBinder.bind(this), and PreferenceBinder.unbind(this)
(if there is at least one annotation without `listen=false`).

This serves to shorten the length of time between the user
forgetting the method call and becoming aware of the issue.

Fixes #19