Fix for NullPointerException during build.
This was occurring when analysing the code for the bind/unbind
calls. If a method in a target class was abstract, it would have a
null method body, which was not properly checked.

A null check has been added to prevent the crash. A check for
abstract class types has also been added, since a bind/unbind
call is unnecessary for abstract classes.

Fixes #32